options: split creators into separate classes
